# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The uniform error body for every error this codebase raises (186 D6).\n\nFastAPI's own ``RequestValidationError`` is the one documented exception:\nit keeps its native ``detail`` list, which carries per-field ``loc``/``msg``\nthat a flattened string would lose.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Granularity": {
        "description": "Canonical granularity tokens for OHLCV data requests.\n\nSee :class:`CycleGranularity` for the daemon/bookkeeping vocabulary; these\nare request tokens and the two are not interchangeable.",
        "enum": [
          "1m",
          "5m",
          "15m",
          "1h",
          "4h",
          "1d",
          "1w",
          "1mo",
          "1q"
        ],
        "title": "Granularity",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Data serving API for OHLCV bars, symbol metadata, gap status, the Kalshi prediction-market catalog and time series, and operations coverage: pass state, source freshness and provider credits.",
    "title": "Manta Trading API",
    "version": "0.16.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/bars/{symbol}": {
      "get": {
        "description": "Return OHLCV bars for ``symbol`` over the requested date range.\n\nFor a cagg-served granularity the response carries ``is_stale``, probed\nagainst the exact view the bars came from (slice 185 D7). ``M1``/``D1`` read\nraw hypertables, so no probe is issued and ``is_stale`` is always ``False``.\n\nWindows are admitted before any database work (186 D4): a reversed range or\none whose estimated bar count exceeds the configured ceiling is a ``422``,\nnot a slow ``500``.",
        "operationId": "get_bars_api_v1_bars__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Granularity"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "adjusted",
            "required": false,
            "schema": {
              "default": true,
              "title": "Adjusted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "msgpack"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "Get Bars"
      }
    },
...
  }
}